Pagination updated
Was able to convert to component
